backend/requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
openai==0.27.0  
flask_cors==3.0.10
gunicorn==20.1.0
